adding more triggers on one event
